Squashed commit of the following:
commit 9d519bb032111b2f2f7dc1f2d188e0e1981f3970
Author: Satyam Ghodasara <sghodas@gmail.com>
Date:   Fri Mar 30 17:23:44 2018 -0400

    Added alert changes to the change log

commit 7499bd3ed00d1c36619a0e051eef33600f8dac9b
Author: Satyam Ghodasara <sghodas@gmail.com>
Date:   Fri Mar 30 17:13:28 2018 -0400

    Updated project

commit e71f5b007482e3581895a25c2261351e86431a5e
Author: Satyam Ghodasara <sghodas@gmail.com>
Date:   Fri Mar 30 17:13:14 2018 -0400

    Bumped version to 4.1.0

commit 4f4a01a39add2495d5186376ea22efe44057249a
Author: Satyam Ghodasara <sghodas@gmail.com>
Date:   Fri Mar 30 17:08:09 2018 -0400

    Updated change log

commit 28a433e2a3e8650b019cc9b663a7801e21164be2
Author: Satyam Ghodasara <sghodas@gmail.com>
Date:   Sun Feb 12 14:48:58 2017 -0500

    Added unit tests for new Alert properties

commit 3cb5fc7b0f2df3903bd439269b58eed4e4b814d9
Author: Satyam Ghodasara <sghodas@gmail.com>
Date:   Sun Feb 12 14:31:06 2017 -0500

    Added time, regions, and severity properties and Severity enum to the Alert struct

commit b20f414a1f383fe1f3f7d7ceea58cc323d1c0935
Author: Satyam Ghodasara <sghodas@gmail.com>
Date:   Sun Feb 12 13:51:12 2017 -0500

    Added unit tests for Bulgarian language

+++ Source/Forecast.swift
@@ -66,8 +66,8 @@ public struct Forecast {
     ///
     /// - returns: A new `Forecast` filled with data from the given JSON object.
     public init(fromJSON json: NSDictionary) {
-        latitude = json["latitude"] as! Float
-        longitude = json["longitude"] as! Float
+        latitude = json["latitude"] as! Double
+        longitude = json["longitude"] as! Double
         timezone = json["timezone"] as! String
         
         if let jsonCurrently = json["currently"] as? NSDictionary {
